+++ pysatNASA/instruments/icon_euv.py
@@ -45,6 +45,7 @@ Russell Stoneback, Mar 23, 2018, University of Texas at Dallas
 
 import datetime as dt
 import functools
+import numpy as np
 
 import pysat
 from pysat.instruments.methods import general as mm_gen
@@ -157,6 +158,31 @@ list_remote_files = functools.partial(cdw.list_remote_files,
                                       supported_tags=download_tags)
 
 
+def filter_metadata(meta_dict):
+    """Filter EUV metadata to remove warnings during loading.
+
+    Parameters
+    ----------
+    meta_dict : dict
+        Dictionary of metadata from file.
+
+    Returns
+    -------
+    dict
+        Filtered EUV metadata.
+
+    """
+    vars = ['ICON_L26_Ancillary_Filename', 'ICON_L26_Flag_Details',
+            'ICON_L26_Input_Filename', 'ICON_L26_Processing_Date',
+            'ICON_L26_UTC_Time']
+
+    for var in vars:
+        if var in meta_dict:
+            meta_dict[var]['FillVal'] = np.nan
+
+    return meta_dict
+
+
 def load(fnames, tag=None, inst_id=None, keep_original_names=False):
     """Load ICON EUV data into `xarray.Dataset` object and `pysat.Meta` objects.
 
@@ -206,9 +232,12 @@ def load(fnames, tag=None, inst_id=None, keep_original_names=False):
               'min_val': ('ValidMin', float),
               'max_val': ('ValidMax', float), 'fill_val': ('FillVal', float)}
 
-    data, meta = pysat.utils.load_netcdf4(fnames, epoch_name='Epoch',
-                                          pandas_format=pandas_format,
-                                          labels=labels)
+    data, meta = pysat.utils.io.load_netcdf(fnames, epoch_name='Epoch',
+                                            pandas_format=pandas_format,
+                                            labels=labels,
+                                            meta_processor=filter_metadata,
+                                            drop_meta_labels=['Valid_Max',
+                                                              'Valid_Min'])
 
     # xarray can't merge if variable and dim names are the same
     if 'Altitude' in data.dims: